--- mission-racing/src/main/java/com/racing/domain/NumberGenerator.java
@@ -1,5 +1,0 @@
-package com.racing.domain;
-
-public interface NumberGenerator {
-    int generateRandom();
-}
