--- mission-racing/src/main/java/com/racing/domain/RandomNumbeGenerate.java
@@ -1,9 +1,0 @@
-package com.racing.domain;
-
-
-public class RandomNumbeGenerate implements NumberGenerator {
-    @Override
-    public int generateRandom() {
-        return (int) (Math.random() * 10);
-    }
-}
