+++ mission-racing/src/racingSpring/src/main/java/com/example/racingspring/domain/NumberGenerator.java
@@ -1,0 +1,5 @@
+package com.example.racingspring.domain;
+
+public interface NumberGenerator {
+    int generateRandom();
+}
