+++ mission-racing/src/racingSpring/src/main/java/com/example/racingspring/domain/RandomNumbeGenerate.java
@@ -1,0 +1,8 @@
+package com.example.racingspring.domain;
+
+public class RandomNumbeGenerate implements NumberGenerator {
+    @Override
+    public int generateRandom() {
+        return (int) (Math.random() * 10);
+    }
+}
