+++ mission-racing/src/racingSpring/src/main/java/com/example/racingspring/domain/도메인로직.md
