+++ mission-racing/src/racingSpring/src/main/java/com/example/racingspring/exception/NullInputException.java
@@ -1,0 +1,8 @@
+package com.example.racingspring.exception;
+
+public class NullInputException extends RuntimeException {
+
+    public NullInputException() {
+        super("자동차 이름을 꼭 입력하세요");
+    }
+}
