+++ mission-racing/src/racingSpring/src/test/java/com/example/racingspring/domain/FakeNumberGenerator.java
@@ -1,0 +1,8 @@
+package com.example.racingspring.domain;
+
+public class FakeNumberGenerator implements NumberGenerator {
+    @Override
+    public int generateRandom() {
+        return 5;
+    }
+}
