--- mission-racing/src/study/핵심애노테이션.md
@@ -1,116 +1,0 @@
-
-# Spring annotation 을 공부해보자 ..!
-
-# 핵심 애노테이션
-
-1. @SpringBootApplication 
-   - Spring boot 애플리케이션의 진입점으로 사용 
-   - @Configuration, @EnableAutoConfiguration, @ComponentScane을 포함한다
-
-2. @Component:
-    - 자동으로 빈으로 등록되도록 지정
-    - 주로 서비스나 레포지토리 클래스에 사용된다.
-    - Spring 컨테이너가 이 클래스를 빈으로 관리하게 한다. 
-
-3. @Service 
-   - @Component 와 동일 서비스 계층에 사용되는 클래스에 의미를 부여한다.
-   - 비즈니스 로직을 처리하는 서비스 계층을 명확히한다. 
-   - 구분 짓기 위해 사용하는 느낌이다!
-
-
-4. @Repository
-   - @Component와 동일하지만, 데이터 액세스 계층에 사용되는 클래스에 의미를 부여하고 
-   - 데이터 엑세스 예외로 변환한다.
-   - 데이터베이스 예외를 스프링의 데이터 액세스 예외로 변환한다.
-
-
-
-5. @Controller 
-   - MVC에서 컨트롤러 역할을 하는 클래스에 사용한다. 
-   -  웹 요청을 처리하고 뷰를 반환하는 컨트롤러로 동작하는듯
-
-6. @RestController
-   - Controller와 @ResponseBody의 조합으로 
-   - RESTful 웹 서비스 컨트롤러에 사용된다. 
-   - JSON/XML 형식의 데이터를 반환하는? 
-
-7. @Configuration: 
-   - 자바 기반 설정 파일을 나타낸다. 빈 정의를 포함할 수있다.
-
-8. @Bean :
-   - 메서드 레벨에서 사용되어 반환된 객체를 Spring의 빈으로 등록
-
-
-# 의존성 주입
-
-1. @Autowired :
-   - 의존성 주입을 위한 애노테이션 
-   - 생성자, 필드, 메서드에 적용
-   
-2. @Qualifier:
-   - @Autowired와 같이 사용 
-   - 명시적으로 주입할 때 사용 
-   - EX) @Qualifier("mySpecificService")
-   - 주입할 빈을 이름으로 명시할 수 있음 코드 명확성을 위한 것인가?
-
-3. @inject 
-   - javax.inject 패키지에서 제공 
-   - @Autowired 와 비슷
-
-4. @Resource 
-    - @Autowired 와 비슷함
-    - @Resource(name = "myService")
-      private MyService myService;
-    - 이런식으로 빈의 이름을 사용하여 주입
-
-
-# Spring MVC 애노테이션 :
-
-
-1. @RequestMapping 
-   - URL 요청을 처리할 메서드나 클래스에 매핑한다. 
-   - 지정된 경로에 대한 요청을 해당 메서드로 매핑
-
-2. @GetMapping @PostMapping @PutMapping, @DeleteMapping, @PatchMapping
-
-    - HTTP 메서드와 URL을 매핑한다.
-    - @RequestMapping 의 축약형
-
-3. @PathVariable
-   - URL 경로에서 변수를추출하여 메서드 파라미터에 바인딩한다. 
-   - URL 경로의 변수를 메서드 파라미터로 추출한다.
-
-4. @RequestParam
-   - 요청 파라미터를 메서드 파라미터에 바인딩한다.
-
-5. @RequestBody
-   - 요청 본문을 메서드 파라미어테 바인딩한다.
-   
-6. @ResponseBody
-   - 메서드의 반환 값을 HTTP 응답 본문으로 변환합니다.(클라이언트에게)
-
-# Spring Data JPA 애노테이션
-
-
-1. @Entity
-   - JPA 엔티티 클래스를 지정 
-
-
-2. @Table 
-   - 엔티티와 매핑할 데이터베이스 테이블을 지정한다.
-   
-
-3. @Id
-
-
-
-
-빈으로 관리한다는 것이 무엇일까? 에 대한 공부
-
-1. 객체의 생명 주기를 컨테이너가 관리하고 
-2. 필요할 때 객체를 주입하거나 사용할 수 있도록 해준다
-3. 의존성 주입을 통해 이루어진다. 
-4. 개발자가 객체의 생성의 관여하지않고 Spring이 해주는 것이다.
-5. 빈으로 등록된 객체는 애플리케이션 전역에서 재사용가능 싱글턴 패턴으로 관리
-
-
