--- mission-racing/src/test/java/com/racing/domain/FakeNumberGenerator.java
@@ -1,8 +1,0 @@
-package com.racing.domain;
-
-public class FakeNumberGenerator implements NumberGenerator {
-    @Override
-    public int generateRandom() {
-        return 5;
-    }
-}
